+++ src/main/java/org/wiremock/extensions/state/internal/model/ContextTemplateModel.java
@@ -28,8 +28,7 @@ public final class ContextTemplateModel {
         return new ContextTemplateModel(context);
     }
 
-
-    public Long getUpdateCount() {
+    public Integer getUpdateCount() {
         return context.getUpdateCount();
     }
 }
